Refactor: ADded loader file path for buyers, and fixed json tag for Last Name in buyer model
