+++ internal/loader/buyerLoader/buyer_json.go
@@ -47,6 +47,5 @@ func (l *JsonFile) Load() (v map[int]models.Buyer, err error) {
 			},
 		}
 	}
-
 	return
 }
