+++ pkg/models/buyer.go
@@ -15,5 +15,5 @@ type BuyerDoc struct {
 	Id           int    `json:"id"`
 	CardNumberId string `json:"card_number_id"`
 	FirstName    string `json:"first_name"`
-	LastName     string `json:"Last_name"`
+	LastName     string `json:"last_name"`
 }
